python/pyproject.toml
[tool.poetry]
name = "arg-services"
version = "0.1.14"
description = "GraphQL and gRPC schemas for Microservice-based Argumentation Machines."
authors = ["Mirko Lenz <info@mirko-lenz.de>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "http://recap.uni-trier.de"
repository = "https://github.com/ReCAP-UTR/Schema"
include = ["arg_services/**/*.py", "arg_services/**/*.pyi"]
packages = [{ include = "arg_services" }, { include = "arg_services_helper" }]

[tool.poetry.dependencies]
python = "^3.6"
grpcio = "^1.32"
protobuf = "^3.15.8"
aiomultiprocess = { version = "^0.9.0", optional = true }
grpcio-reflection = "^1.39.0"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.32"
mypy-protobuf = "^2.4"
mypy = "^0.812"

[tool.poetry.extras]
server = ["aiomultiprocess"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
